Move files to folders following MVC structure
